requirements.txt
Flask==2.3.3
gunicorn==20.1.0
ultralytics==8.3.59
pillow==11.1.0
torch==2.1.0
torchvision==0.16.0
numpy==1.23.0
opencv-python==4.10.0.84
requests==2.32.3
boto3
